+++ plugin/download.sh
@@ -34,22 +34,20 @@ download() {
 
 delete_old_plugin() {
     local dir=${1}
-    rm -f "$dir/heroku-nodejs-plugin-node-12.sha512"
-    rm -f "$dir/heroku-nodejs-plugin-node-12.tar.gz"
     rm -f "$dir/heroku-nodejs-plugin-node-14.sha512"
     rm -f "$dir/heroku-nodejs-plugin-node-14.tar.gz"
     rm -f "$dir/heroku-nodejs-plugin-node-16.sha512"
     rm -f "$dir/heroku-nodejs-plugin-node-16.tar.gz"
+    rm -f "$dir/heroku-nodejs-plugin-node-17.sha512"
+    rm -f "$dir/heroku-nodejs-plugin-node-17.tar.gz"
+    rm -f "$dir/heroku-nodejs-plugin-node-18.sha512"
+    rm -f "$dir/heroku-nodejs-plugin-node-18.tar.gz"
     rm -f "$dir/version"
 }
 
 download_assets_for_release() {
     local tag=${1}
     local dir=${2}
-
-    # Node 12
-    download "https://github.com/heroku/heroku-nodejs-plugin/releases/download/$tag/heroku-nodejs-plugin-node-12-$tag.sha512" "$dir/heroku-nodejs-plugin-node-12.sha512"
-    download "https://github.com/heroku/heroku-nodejs-plugin/releases/download/$tag/heroku-nodejs-plugin-node-12-$tag.tar.gz" "$dir/heroku-nodejs-plugin-node-12.tar.gz"
 
     # Node 14
     download "https://github.com/heroku/heroku-nodejs-plugin/releases/download/$tag/heroku-nodejs-plugin-node-14-$tag.sha512" "$dir/heroku-nodejs-plugin-node-14.sha512"
@@ -58,6 +56,14 @@ download_assets_for_release() {
     # Node 16
     download "https://github.com/heroku/heroku-nodejs-plugin/releases/download/$tag/heroku-nodejs-plugin-node-16-$tag.sha512" "$dir/heroku-nodejs-plugin-node-16.sha512"
     download "https://github.com/heroku/heroku-nodejs-plugin/releases/download/$tag/heroku-nodejs-plugin-node-16-$tag.tar.gz" "$dir/heroku-nodejs-plugin-node-16.tar.gz"
+
+    # Node 17
+    download "https://github.com/heroku/heroku-nodejs-plugin/releases/download/$tag/heroku-nodejs-plugin-node-17-$tag.sha512" "$dir/heroku-nodejs-plugin-node-17.sha512"
+    download "https://github.com/heroku/heroku-nodejs-plugin/releases/download/$tag/heroku-nodejs-plugin-node-17-$tag.tar.gz" "$dir/heroku-nodejs-plugin-node-17.tar.gz"
+
+    # Node 18
+    download "https://github.com/heroku/heroku-nodejs-plugin/releases/download/$tag/heroku-nodejs-plugin-node-18-$tag.sha512" "$dir/heroku-nodejs-plugin-node-18.sha512"
+    download "https://github.com/heroku/heroku-nodejs-plugin/releases/download/$tag/heroku-nodejs-plugin-node-18-$tag.tar.gz" "$dir/heroku-nodejs-plugin-node-18.tar.gz"
 }
 
 test_hash() {
@@ -90,8 +96,9 @@ echo $TAG_NAME > "$PLUGIN_DIR/version"
 
 echo "Plugins downloaded"
 
-test_hash 12 $PLUGIN_DIR
 test_hash 14 $PLUGIN_DIR
 test_hash 16 $PLUGIN_DIR
+test_hash 17 $PLUGIN_DIR
+test_hash 18 $PLUGIN_DIR
 
 echo "Done"